@voidxno 2023-11-28
Rev02 commit reworked to achieve logic commented above.