@voidxno 2023-11-29
**[DONE]** Testing with Rev03 commit:
- Compiled and ran with Linux/gcc, Linux/Clang, and Win10/VC++
- Logic of tests:
  - Live mmx-node, disable GPU, only CPU SHA-NI verify
  - Live mmx-node has fixed peer with my regular mmx-node
  - Look at logging both mmx-node's for any verify warnings/peers

**[DONE]** Separate test with Rev03 (w/custom debug logging):
- Ran live mmx-node on Linux/gcc and waited for:
  - Odd number `num_lanes` (observed, ok verify)
  - Different `num_iters` two points (observed, ok verify)

**[DONE]** Improved CPU verify VDF times still ok (checked, Rev03).

**Ready for merge from my side.**